Moved generation of new rumors to server
